Restructured SoundFont.elm
Got rid of the global state in the javascript implementation which was a global AudioContext.  Instead, introduced a function getAudioContext and then pass this to each function that requires it.  Added a function isWebAudioEnabled so clients can detect whether or not the browser will support playing WebAudio.

Altered the timings for the Type0 player so that tunes play at the specified speed by introducing an 'elm slowdown factor'.